enhanced side bar components test
