+++ src/test/skyboxInfo.test.tsx
@@ -1,12 +1,9 @@
 import { render } from "@testing-library/react";
 import React from "react";
-import CategoryIcon from "@mui/icons-material/Category";
-import AssetItem from "../components/Editor/TreeView/AssetItem";
+import SkyBoxInfo from "../components/Editor/SideBar/SkyBoxInfo";
 
-test("ActionItem renders without crashing", () => {
-  render(<AssetItem labelText="TestItem" nodeId="Test" LabelIcon={CategoryIcon} />);
+test("Skybox renders without crashing", () => {
+  render(<SkyBoxInfo/>);
 });
 
-test("ActionItem renders without crashing with undefined LabelText", () => {
-  render(<AssetItem labelText={undefined} nodeId="Test" LabelIcon={CategoryIcon} />);
-});
+
