+++ src/test/textEditor.test.tsx
@@ -1,9 +1,14 @@
-import { render } from "@testing-library/react";
+import { fireEvent, render, screen } from "@testing-library/react";
 import React from "react";
-import SkyBoxInfo from "../components/Editor/SideBar/SkyBoxInfo";
+import TextEditor from "../components/Editor/TextEditor";
+import { ActionContext } from "../util/customHooks/actionContext";
 
-test("ActionItem renders without crashing", () => {
-  render(<SkyBoxInfo/>);
+test("Text editor renders without crashing", () => {
+  render(
+    <ActionContext.Provider value={{ addActionToList: () => undefined }}>
+      <TextEditor />
+    </ActionContext.Provider>
+  );
+  const button = screen.getByText("ADD/EDIT DESCRIPTION");
+  fireEvent.click(button);
 });
-
-
